refactor: sse import 경로 변경
